added refund and splitted test into different files
