+++ test/setup.ts
@@ -4,11 +4,11 @@ import { RIFScheduler__factory } from '@rsksmart/rif-scheduler-contracts/dist/et
 import { ERC677__factory } from './contracts/types/factories/ERC677__factory'
 import ERC677Data from './contracts/ERC677.json'
 import { utils, Signer, BigNumber, constants, providers } from 'ethers'
-import { JsonRpcProvider } from '@ethersproject/providers'
 import { IPlanResponse } from '../src'
 
 const Config = {
-  BLOCKCHAIN_HTTP_URL: 'HTTP://127.0.0.1:8545'
+  BLOCKCHAIN_HTTP_URL: 'HTTP://127.0.0.1:8545',
+  MINIMUM_TIME_BEFORE_EXECUTION: 16 // sec
 }
 
 const getJsonRpcProvider = async function (): Promise<providers.JsonRpcProvider> {
@@ -35,7 +35,7 @@ const getUsers = async function ():Promise<users> {
 const contractsSetUp = async function (): Promise<{schedulerAddress:string, tokenAddress:string, tokenAddress677:string}> {
   utils.Logger.setLogLevel(utils.Logger.levels.OFF)
   const users = await getUsers()
-  const oneShotScheduleFactory = new RIFScheduler__factory(users.admin)
+  const rifSchedulerFactory = new RIFScheduler__factory(users.admin)
 
   const erc677Factory = new ERC677__factory(users.admin)
   const erc677 = await erc677Factory.deploy(await users.admin.getAddress(), BigNumber.from(100000), 'RIF', 'RIF')
@@ -45,8 +45,11 @@ const contractsSetUp = async function (): Promise<{schedulerAddress:string, toke
   const erc20 = await erc677Factory.deploy(await users.admin.getAddress(), BigNumber.from(100000), 'DOC', 'DOC')
   await erc20.transfer(await users.serviceConsumer.getAddress(), BigNumber.from(50000))
 
-  const schedulerContract = await oneShotScheduleFactory.deploy(await users.serviceProvider.getAddress(), await users.payee.getAddress())
-  // ethers.utils.Logger.setLogLevel(ethers.utils.Logger.levels.WARNING)
+  const schedulerContract = await rifSchedulerFactory.deploy(
+    await users.serviceProvider.getAddress(),
+    await users.payee.getAddress(),
+    Config.MINIMUM_TIME_BEFORE_EXECUTION
+  )
   return { schedulerAddress: schedulerContract.address, tokenAddress: erc20.address, tokenAddress677: erc677.address }
 }
